feat(time): users can update time records

+++ time/src/components/ProjectTimeRecords.tsx
@@ -1,9 +1,9 @@
-import { Action, ActionPanel, Icon, Image, List } from "@raycast/api";
+import { Action, ActionPanel, Alert, confirmAlert, Icon, Image, List } from "@raycast/api";
 import { orderBy, sum } from "es-toolkit";
 import { DateTime } from "luxon";
 import { useTimeClient } from "../hooks/useTimeClient";
 import { Project } from "../types/Project";
-import { CreateTimeRecord } from "./CreateTimeRecord";
+import { CreateUpdateTimeRecord } from "./CreateUpdateTimeRecord";
 
 type ProjectTimeRecordsProps = {
   project: Project;
@@ -17,12 +17,34 @@ export const ProjectTimeRecords: React.FC<ProjectTimeRecordsProps> = ({ project 
   const totalHours = sum(orderedTimeRecords.map((timeRecord) => timeRecord.duration)) / 100;
 
   const onDeleteTimeRecord = async (id: number) => {
-    await deleteTimeRecord(id);
-    revalidate();
+    if (
+      await confirmAlert({
+        title: "Delete Time Record",
+        message: "Are you sure you want to delete this time record?",
+        primaryAction: { title: "Delete", style: Alert.ActionStyle.Destructive },
+      })
+    ) {
+      await deleteTimeRecord(id);
+      revalidate();
+    }
   };
 
   return (
     <List isLoading={isLoading} navigationTitle={`Time records -> ${project.name}`}>
+      {orderedTimeRecords.length === 0 && (
+        <List.EmptyView
+          title="No time records found for this project."
+          actions={
+            <ActionPanel>
+              <Action.Push
+                title="Create Time Record"
+                icon={Icon.Plus}
+                target={<CreateUpdateTimeRecord project={project} />}
+              />
+            </ActionPanel>
+          }
+        />
+      )}
       <List.Section title={`${DateTime.now().toFormat("MMMM")} -> ${totalHours.toFixed(1)}h`}>
         {orderedTimeRecords.map((timeRecord) => (
           <List.Item
@@ -40,14 +62,27 @@ export const ProjectTimeRecords: React.FC<ProjectTimeRecordsProps> = ({ project 
             actions={
               <ActionPanel>
                 <Action.Push
-                  title="Create Time Record"
-                  icon={Icon.Plus}
-                  target={<CreateTimeRecord project={project} onTimeRecordCreated={revalidate} />}
+                  title="Update Time Record"
+                  icon={Icon.Pencil}
+                  target={
+                    <CreateUpdateTimeRecord
+                      timeRecord={timeRecord}
+                      project={project}
+                      onTimeRecordCreated={revalidate}
+                    />
+                  }
                 />
                 <Action
+                  style={Action.Style.Destructive}
+                  shortcut={{ modifiers: ["cmd"], key: "backspace" }}
                   title="Delete Time Record"
                   icon={Icon.Trash}
                   onAction={() => onDeleteTimeRecord(timeRecord.id)}
+                />
+                <Action.Push
+                  title="Create Time Record"
+                  icon={Icon.Plus}
+                  target={<CreateUpdateTimeRecord project={project} />}
                 />
               </ActionPanel>
             }
